example of bar?asdf=1 -> foo?asdf=1 redirect.
To preserve queryParams, they should be persistent on the controller in question. As bar controller does not specify `asdf` as a QP, when the transition occurs, `asdf` is assumed to be null and removed.

I don’t really like this design, I think it should be improved/changed. And when we (this year) re-work query params we should take this into account